+++ src/main/java/com/example/smartmessaging/dto/vo/UsersVO.java
@@ -5,7 +5,6 @@ import lombok.Setter;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -13,8 +12,8 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @Builder
 public class UsersVO extends BaseVO {
-    private BigDecimal userId;
-    private BigDecimal empNum;
+    private Long userId;
+    private Integer empNum;
     private String pwd;
     private String name;
     private Boolean isActive;
